Init virtual tree refactor
This refactor is intended to overcome certain limitations in the current
architecture, where there’s a lot of situations where we would need to
"seek in the future" to get the right value.

In this new refactor, the goal is to incrementally build a tree
virtually representing the final repo, with multiple passes.
Then, when finished, another module will walk the tree, display in
terminal and write the files.